feat(referral relation):

+++ aspnetcore/src/Crm.Domain/Referrals/IReferralRelationRepository.cs
@@ -9,8 +9,8 @@ namespace Crm.Referrals;
 public interface
     IReferralRelationRepository : IAstraBasicWithPagedBasicRepository<ReferralRelation, Guid, ReferralRelationPagedParameter>
 {
-    Task<List<ReferralRelation>> GetAncestorRelationListAsync(Guid recommendeeId, ushort? minDepth = null);
-    Task<List<ReferralRelation>> GetDescendantRelationListAsync(Guid ancestorId, ushort? minDepth = null);
+    Task<List<ReferralRelation>> GetAncestorRelationListAsync(Guid recommendeeId, uint? minDepth = null);
+    Task<List<ReferralRelation>> GetDescendantRelationListAsync(Guid ancestorId, uint? minDepth = null);
     Task<ReferralRelation?> FindParentAsync(Guid recommendeeId);
     Task<bool> ExistsAsync(Guid recommendeeId);
     Task<PagedList<RecommendeeQueryModel>> GetRecommendeePagedListAsync(RecommendeeQueryModelPagedParameter parameter);
